+++ src/test/java/org/folio/entlinks/service/messaging/authority/InstanceAuthorityLinkUpdateServiceTest.java
@@ -2,8 +2,10 @@ package org.folio.entlinks.service.messaging.authority;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.support.base.TestConstants.TENANT_ID;
+import static org.folio.support.base.TestConstants.USER_ID;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.never;
@@ -224,14 +226,15 @@ class InstanceAuthorityLinkUpdateServiceTest {
     when(sourceRecordService.getAuthoritySourceRecordById(any())).thenReturn(sourceRecord);
     when(updateHandler.handle(changeHolderCaptor.capture())).thenReturn(List.of(expected));
     when(folioExecutionContext.getTenantId()).thenReturn(TENANT_ID);
+    when(folioExecutionContext.getUserId()).thenReturn(UUID.fromString(USER_ID));
     when(consortiumTenantsService.getConsortiumTenants(TENANT_ID)).thenReturn(memberTenants);
     mockExecutionService();
 
     service.handleAuthoritiesChanges(authorityEvents);
 
     verify(eventProducer, times(3)).sendMessages(eventCaptor.capture());
-    verify(executionService).executeSystemUserScoped(eq(memberTenants.get(0)), any());
-    verify(executionService).executeSystemUserScoped(eq(memberTenants.get(1)), any());
+    verify(executionService).executeSystemUserScoped(eq(memberTenants.get(0)), eq(USER_ID), any());
+    verify(executionService).executeSystemUserScoped(eq(memberTenants.get(1)), eq(USER_ID), any());
 
     var messages = eventCaptor.getAllValues().stream().flatMap(Collection::stream).toList();
     assertThat(messages).hasSize(3);
@@ -249,7 +252,7 @@ class InstanceAuthorityLinkUpdateServiceTest {
 
   @SuppressWarnings("unchecked")
   private void mockExecutionService() {
-    doAnswer(invocationOnMock -> ((Callable<Object>) invocationOnMock.getArgument(1)).call())
-      .when(executionService).executeSystemUserScoped(any(), any());
+    doAnswer(invocationOnMock -> ((Callable<Object>) invocationOnMock.getArgument(2)).call())
+      .when(executionService).executeSystemUserScoped(any(), anyString(), any());
   }
 }